refactor: 인터페이스로 구성
