+++ src/main/java/com/coupang/marketplace/global/util/login/HttpSessionUtil.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.global.util.session;
+package com.coupang.marketplace.global.util.login;
 
 import javax.servlet.http.HttpSession;
 
@@ -10,18 +10,21 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Component
-public class HttpSessionUtil {
+public class HttpSessionUtil implements LoginUtil{
 
 	private final HttpSession httpSession;
 
+	@Override
 	public Object getAttribute(){
 		return httpSession.getAttribute(SessionKey.LOGIN_USER_ID);
 	}
 
+	@Override
 	public void setAttribute(Object value){
 		httpSession.setAttribute(SessionKey.LOGIN_USER_ID, value);
 	}
 
+	@Override
 	public void removeAttribute(){
 		httpSession.removeAttribute(SessionKey.LOGIN_USER_ID);
 	}
