+++ src/main/java/com/coupang/marketplace/global/util/login/LoginUtil.java
@@ -1,0 +1,8 @@
+package com.coupang.marketplace.global.util.login;
+
+public interface LoginUtil {
+
+	Object getAttribute();
+	void setAttribute(Object value);
+	void removeAttribute();
+}
